refactor: вынести все url в конфиг
