+++ src/constants/config.js
@@ -1,0 +1,3 @@
+export const URL_NOMOREPARTIES = "https://api.nomoreparties.co";
+export const URL_MOVIES = `${URL_NOMOREPARTIES}/beatfilm-movies`;
+export const URL_MAIN = "https://api.movies-explore.nomoredomains.xyz";
