--- src/constants/constants.js
@@ -1,2 +1,0 @@
-export const URL_MOVIES = "https://api.nomoreparties.co/beatfilm-movies";
-export const URL_MAIN = "https://api.movies-explore.nomoredomains.xyz";
