+++ src/utils/MainApi.js
@@ -1,4 +1,5 @@
-import { URL_MAIN } from "../constants/constants";
+import { URL_MAIN } from "../constants/config";
+import { URL_NOMOREPARTIES } from "../constants/config";
 import { createHeaders } from "./headers";
 
 class MainApi {
@@ -64,9 +65,9 @@ class MainApi {
         duration: movie.duration,
         year: movie.year,
         description: movie.description,
-        image: `https://api.nomoreparties.co${movie.image.url}`,
+        image: `${URL_NOMOREPARTIES}${movie.image.url}`,
         trailerLink: movie.trailerLink,
-        thumbnail: `https://api.nomoreparties.co${movie.image.formats.thumbnail.url}`,
+        thumbnail: `${URL_NOMOREPARTIES}${movie.image.formats.thumbnail.url}`,
         movieId: movie.id,
         nameRU: movie.nameRU,
         nameEN: movie.nameEN,
